Simplify project annotation config dirs
