+++ src/components/Text/Annotated/Annotated.tsx
@@ -1,6 +1,6 @@
 import { BroccoliTextGeneric } from "../../../model/Broccoli.ts";
 import { useAnnotationStore } from "../../../stores/annotation.ts";
-import "../annotated.css";
+import "./annotated.css";
 import {
   projectConfigSelector,
   useProjectStore,
