+++ src/projects/default/config/index.tsx
@@ -12,7 +12,7 @@ import { getCategory } from "./getCategory.ts";
 import { isEntity } from "./isEntity.ts";
 import { getTocId, showToc } from "./showToc.ts";
 import { getUrl, isLink } from "./isLink.ts";
-import { defaultAnnotatedTextConfig } from "../defaultAnnotatedTextConfig.ts";
+import { defaultAnnotatedTextConfig } from "../annotation/defaultAnnotatedTextConfig.ts";
 
 /**
  * Default configuration file with some sensible defaults
