+++ src/projects/kunstenaarsbrieven/annotation/marker/InsertMarkerAnnotation.tsx
@@ -1,6 +1,6 @@
 import { MarkerSegment } from "../../../../components/Text/Annotated/core";
 import { MarkerBody } from "../../../../model/AnnoRepoAnnotation.ts";
-import { isHeadBody } from "../../annotation/ProjectAnnotationModel.ts";
+import { isHeadBody } from "../ProjectAnnotationModel.ts";
 
 type InsertMarkerAnnotationProps = {
   marker: MarkerSegment<MarkerBody>;
